+++ moabb/datasets/fake.py
@@ -20,9 +20,11 @@ class FakeDataset(BaseDataset):
         n_runs=2,
         n_subjects=10,
         paradigm="imagery",
+        channels=("C3", "Cz", "C4"),
     ):
         self.n_runs = n_runs
         event_id = {ev: ii + 1 for ii, ev in enumerate(event_list)}
+        self.channels = channels
         super().__init__(
             list(range(1, n_subjects + 1)),
             n_sessions,
@@ -42,21 +44,18 @@ class FakeDataset(BaseDataset):
         return data
 
     def _generate_raw(self):
-
-        ch_names = ["C3", "Cz", "C4"]
-
         montage = make_standard_montage("standard_1005")
         sfreq = 128
         duration = len(self.event_id) * 60
-        eeg_data = 2e-5 * np.random.randn(duration * sfreq, len(ch_names))
+        eeg_data = 2e-5 * np.random.randn(duration * sfreq, len(self.channels))
         y = np.zeros((duration * sfreq))
         for ii, ev in enumerate(self.event_id):
             start_idx = (1 + 5 * ii) * 128
             jump = 5 * len(self.event_id) * 128
             y[start_idx::jump] = self.event_id[ev]
 
-        ch_types = ["eeg"] * len(ch_names) + ["stim"]
-        ch_names = ch_names + ["stim"]
+        ch_types = ["eeg"] * len(self.channels) + ["stim"]
+        ch_names = list(self.channels) + ["stim"]
 
         eeg_data = np.c_[eeg_data, y]